requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
dj-database-url==3.0.1
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==22.0.0